langfuse/pyproject.toml
[project]
name = "review-testcase-langfuse"
version = "0.1.0"
description = "Langfuse интеграция для системы AI-ревью тест-кейсов"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langfuse>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
